envs/gen_env/pyproject.toml
[project]
name = "gen-env"
version = "0.1.0"
description = "Genesis Environment — self-improving coding agent"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "fastmcp>=3.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "anthropic>=0.25.0",
    "rank-bm25>=0.2.2",
    "pyflakes>=3.0.0",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]
